Incomplete changes to the home page for mobile compatibility
Made some changes to the following:
- The more about us button and the text above it.
- "Who we are" div.
- "Work with us and get your site now at no cost" div.

These 3 should (hopefully) be at least fully visible on mobile now. Could not figure out how to make the conversation compatible.

My plan for the main image and the articles at the bottom is to make them stack on top of one another. The main image will go below the text div in MainImage, and the articles at the bottom will be stacked on top of each other. All of this will be complete hopefully by the end of tomorrow.(Monday)